Refaktor: Text-Utilities ausgelagert
