+++ web/src/fileUtils.mjs
@@ -1,0 +1,4 @@
+import { createRequire } from 'module';
+const require = createRequire(import.meta.url);
+const { calculateTextSimilarity, levenshteinDistance } = require('./fileUtils.js');
+export { calculateTextSimilarity, levenshteinDistance };
